+++ packages/atomic/cypress/integration/result-templates/result-link.cypress.ts
@@ -1,0 +1,21 @@
+describe('Result Link Component', () => {
+  describe('when not a child of an "template" element', () => {
+    it.skip('should render an "atomic-component-error" component');
+  });
+
+  it.skip('the "target" prop should set the target on the "a" tag');
+
+  it.skip(
+    'the "href" attribute of the "a" tag should be the result\'s clickUri'
+  );
+
+  describe('when there is a slot', () => {
+    it.skip('should render the slot inside of the "a" tag');
+  });
+
+  describe('when there is not slot', () => {
+    it.skip(
+      'should render an "atomic-result-text" component containing the title'
+    );
+  });
+});