+++ packages/headless/src/controllers/pager/headless-pager.ts
@@ -39,7 +39,7 @@ export interface PagerOptions {
   /**
    * The number of pages to display in the pager.
    *
-   * @default 5
+   * @defaultValue `5`
    * */
   numberOfPages?: number;
 }