+++ packages/quantic/copy-static-resources.js
@@ -15,21 +15,6 @@ const main = async () => {
   );
 
   console.info('Headless copied.');
-
-  await copy(
-    './node_modules/@coveo/atomic/dist',
-    './force-app/main/default/staticresources/atomicutils'
-  );
-
-  console.info('Atomic copied.');
-
-  await copy(
-    './node_modules/tributejs/dist',
-    './force-app/main/default/staticresources/tributejs'
-  );
-
-  console.info('Tributejs copied.');
-  return;
 };
 
 main().then(() => {